[APM] Migrate service tests to deployment agnostic (#199812)
## Summary

Closes https://github.com/elastic/kibana/issues/198988
Part of https://github.com/elastic/kibana/issues/193245

This PR contains the changes to migrate `service` test folder to
Deployment-agnostic testing strategy.

>[!NOTE]
> `top_services.spec.ts` and `throughput.spec.ts` were partially
migrated and `annotations.spec.ts` was not migrated

### How to test

- Serverless

```
node scripts/functional_tests_server --config x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.serverless.config.ts
node scripts/functional_test_runner --config x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.serverless.config.ts --grep="APM"
```

It's recommended to be run against
[MKI](https://github.com/crespocarlos/kibana/blob/main/x-pack/test_serverless/README.md#run-tests-on-mki)

- Stateful
```
node scripts/functional_tests_server --config x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.stateful.config.ts
node scripts/functional_test_runner --config x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.stateful.config.ts --grep="APM"
```

- [ ] ~(OPTIONAL, only if a test has been unskipped) Run flaky test
suite~
- [x] local run for serverless
- [x] local run for stateful
- [x] MKI run for serverless

---------

(cherry picked from commit 2e9926de3047ccd47f686406e4f1308d1c38df73)

# Conflicts:
#	x-pack/test/apm_api_integration/tests/services/agent.spec.ts
