+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/alerts/transaction_duration.spec.ts
@@ -12,7 +12,7 @@ import { apm, timerange } from '@kbn/apm-synthtrace-client';
 import expect from '@kbn/expect';
 import { omit } from 'lodash';
 import type { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
-import type { RoleCredentials, SupertestWithRoleScopeType } from '../../../../services';
+import type { RoleCredentials } from '../../../../services';
 import type { DeploymentAgnosticFtrProviderContext } from '../../../../ftr_provider_context';
 import {
   fetchServiceInventoryAlertCounts,
@@ -24,7 +24,6 @@ import {
 } from './helpers/alerting_helper';
 
 export default function ApiTest({ getService }: DeploymentAgnosticFtrProviderContext) {
-  const roleScopedSupertest = getService('roleScopedSupertest');
   const apmApiClient = getService('apmApi');
   const synthtrace = getService('synthtrace');
   const alertingApi = getService('alertingApi');
@@ -43,18 +42,9 @@ export default function ApiTest({ getService }: DeploymentAgnosticFtrProviderCon
 
   describe('transaction duration alert', () => {
     let apmSynthtraceEsClient: ApmSynthtraceEsClient;
-    let supertestViewerWithCookieCredentials: SupertestWithRoleScopeType;
     let roleAuthc: RoleCredentials;
 
     before(async () => {
-      supertestViewerWithCookieCredentials = await roleScopedSupertest.getSupertestWithRoleScope(
-        'viewer',
-        {
-          withInternalHeaders: true,
-          useCookieHeader: true,
-        }
-      );
-
       roleAuthc = await samlAuth.createM2mApiKeyWithRoleScope('admin');
 
       const opbeansJava = apm
@@ -86,7 +76,6 @@ export default function ApiTest({ getService }: DeploymentAgnosticFtrProviderCon
 
     after(async () => {
       await apmSynthtraceEsClient.clean();
-      await supertestViewerWithCookieCredentials.destroy();
       await samlAuth.invalidateM2mApiKeyWithRoleScope(roleAuthc);
     });
