+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/index.ts
@@ -22,6 +22,7 @@ export default function apmApiIntegrationTests({
     loadTestFile(require.resolve('./correlations'));
     loadTestFile(require.resolve('./entities'));
     loadTestFile(require.resolve('./cold_start'));
+    loadTestFile(require.resolve('./services'));
     loadTestFile(require.resolve('./historical_data'));
     loadTestFile(require.resolve('./observability_overview'));
     loadTestFile(require.resolve('./latency'));
