+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/services/service_icons/generate_data.ts
