+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/services/services_detailed_statistics.spec.ts
@@ -13,54 +13,20 @@ import { ApmDocumentType } from '@kbn/apm-plugin/common/document_type';
 import { RollupInterval } from '@kbn/apm-plugin/common/rollup';
 import { apm, timerange } from '@kbn/apm-synthtrace-client';
 import { uniq, map } from 'lodash';
-import { FtrProviderContext } from '../../common/ftr_provider_context';
+import { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
+import type { DeploymentAgnosticFtrProviderContext } from '../../../../ftr_provider_context';
 
 type ServicesDetailedStatisticsReturn =
   APIReturnType<'POST /internal/apm/services/detailed_statistics'>;
 
-export default function ApiTest({ getService }: FtrProviderContext) {
-  const registry = getService('registry');
-
-  const apmApiClient = getService('apmApiClient');
-
-  const synthtrace = getService('apmSynthtraceEsClient');
+export default function ApiTest({ getService }: DeploymentAgnosticFtrProviderContext) {
+  const apmApiClient = getService('apmApi');
+  const synthtrace = getService('synthtrace');
 
   const start = '2021-01-01T00:00:00.000Z';
   const end = '2021-01-01T00:59:59.999Z';
 
   const serviceNames = ['my-service'];
-
-  registry.when(
-    'Services detailed statistics when data is generated',
-    { config: 'basic', archives: [] },
-    () => {
-      it('handles the empty state', async () => {
-        const response = await apmApiClient.readUser({
-          endpoint: `POST /internal/apm/services/detailed_statistics`,
-          params: {
-            query: {
-              start,
-              end,
-              environment: 'ENVIRONMENT_ALL',
-              kuery: '',
-              probability: 1,
-              documentType: ApmDocumentType.TransactionMetric,
-              rollupInterval: RollupInterval.OneMinute,
-              bucketSizeInSeconds: 60,
-              _inspect: true,
-            },
-            body: {
-              serviceNames: JSON.stringify(serviceNames),
-            },
-          },
-        });
-
-        expect(response.status).to.be(200);
-        expect(response.body.currentPeriod).to.be.empty();
-        expect(response.body.previousPeriod).to.be.empty();
-      });
-    }
-  );
 
   async function getStats(
     overrides?: Partial<
@@ -90,12 +56,38 @@ export default function ApiTest({ getService }: FtrProviderContext) {
     return response.body;
   }
 
-  // FLAKY: https://github.com/elastic/kibana/issues/177511
-  registry.when(
-    'Services detailed statistics when data is generated',
-    { config: 'basic', archives: [] },
-    () => {
+  describe('Services detailed statistics', () => {
+    describe('when data is not generated', () => {
+      it('handles the empty state', async () => {
+        const response = await apmApiClient.readUser({
+          endpoint: `POST /internal/apm/services/detailed_statistics`,
+          params: {
+            query: {
+              start,
+              end,
+              environment: 'ENVIRONMENT_ALL',
+              kuery: '',
+              probability: 1,
+              documentType: ApmDocumentType.TransactionMetric,
+              rollupInterval: RollupInterval.OneMinute,
+              bucketSizeInSeconds: 60,
+              _inspect: true,
+            },
+            body: {
+              serviceNames: JSON.stringify(serviceNames),
+            },
+          },
+        });
+
+        expect(response.status).to.be(200);
+        expect(response.body.currentPeriod).to.be.empty();
+        expect(response.body.previousPeriod).to.be.empty();
+      });
+    });
+
+    describe('when data is generated', () => {
       let servicesDetailedStatistics: ServicesDetailedStatisticsReturn;
+      let apmSynthtraceEsClient: ApmSynthtraceEsClient;
 
       const instance = apm.service('my-service', 'production', 'java').instance('instance');
 
@@ -103,12 +95,28 @@ export default function ApiTest({ getService }: FtrProviderContext) {
       const EXPECTED_LATENCY = 1000;
       const EXPECTED_FAILURE_RATE = 0.25;
 
+      function checkStats() {
+        const stats = servicesDetailedStatistics.currentPeriod['my-service'];
+
+        expect(stats).not.empty();
+
+        expect(uniq(map(stats.throughput, 'y'))).eql([EXPECTED_TPM], 'tpm');
+
+        expect(uniq(map(stats.latency, 'y'))).eql([EXPECTED_LATENCY * 1000], 'latency');
+
+        expect(uniq(map(stats.transactionErrorRate, 'y'))).eql(
+          [EXPECTED_FAILURE_RATE],
+          'errorRate'
+        );
+      }
+
       before(async () => {
+        apmSynthtraceEsClient = await synthtrace.createApmSynthtraceEsClient();
         const interval = timerange(new Date(start).getTime(), new Date(end).getTime() - 1).interval(
           '1m'
         );
 
-        await synthtrace.index([
+        await apmSynthtraceEsClient.index([
           interval.rate(3).generator((timestamp) => {
             return instance
               .transaction('GET /api')
@@ -133,22 +141,7 @@ export default function ApiTest({ getService }: FtrProviderContext) {
         ]);
       });
 
-      after(() => synthtrace.clean());
-
-      function checkStats() {
-        const stats = servicesDetailedStatistics.currentPeriod['my-service'];
-
-        expect(stats).not.empty();
-
-        expect(uniq(map(stats.throughput, 'y'))).eql([EXPECTED_TPM], 'tpm');
-
-        expect(uniq(map(stats.latency, 'y'))).eql([EXPECTED_LATENCY * 1000], 'latency');
-
-        expect(uniq(map(stats.transactionErrorRate, 'y'))).eql(
-          [EXPECTED_FAILURE_RATE],
-          'errorRate'
-        );
-      }
+      after(() => apmSynthtraceEsClient.clean());
 
       describe('and transaction metrics are used', () => {
         before(async () => {
@@ -184,6 +177,6 @@ export default function ApiTest({ getService }: FtrProviderContext) {
           checkStats();
         });
       });
-    }
-  );
+    });
+  });
 }
