+++ x-pack/test/apm_api_integration/tests/services/archive_services_detailed_statistics.spec.ts
@@ -28,38 +28,6 @@ export default function ApiTest({ getService }: FtrProviderContext) {
   const serviceNames = ['opbeans-java', 'opbeans-go'];
 
   registry.when(
-    'Services detailed statistics when data is not loaded',
-    { config: 'basic', archives: [] },
-    () => {
-      it('handles the empty state', async () => {
-        const response = await apmApiClient.readUser({
-          endpoint: `POST /internal/apm/services/detailed_statistics`,
-          params: {
-            query: {
-              start,
-              end,
-              environment: 'ENVIRONMENT_ALL',
-              kuery: '',
-              offset: '1d',
-              probability: 1,
-              documentType: ApmDocumentType.TransactionMetric,
-              rollupInterval: RollupInterval.OneMinute,
-              bucketSizeInSeconds: 60,
-            },
-            body: {
-              serviceNames: JSON.stringify(serviceNames),
-            },
-          },
-        });
-
-        expect(response.status).to.be(200);
-        expect(response.body.currentPeriod).to.be.empty();
-        expect(response.body.previousPeriod).to.be.empty();
-      });
-    }
-  );
-
-  registry.when(
     'Services detailed statistics when data is loaded',
     { config: 'basic', archives: [archiveName] },
     () => {
